refactor: split DataTable
