+++ packages/curve-ui-kit/src/shared/ui/DataTable/index.ts
@@ -1,0 +1,2 @@
+export * from './DataTable'
+export * from './data-table.utils'
